use generalized timer impl
