+++ challenge-manager/challenge-tree/ancestors_test.go
@@ -60,28 +60,28 @@ func TestAncestors_AllChallengeLevels(t *testing.T) {
 
 	t.Run("junk edge fails", func(t *testing.T) {
 		// We start by querying for ancestors for a block edge id.
-		_, _, err := tree.HonestPathTimer(ctx, id("foo"), blockNum)
+		_, err := tree.ComputeAncestorsWithTimers(ctx, id("foo"), blockNum)
 		require.ErrorContains(t, err, "not found in honest challenge tree")
 	})
 	t.Run("dishonest edge lookup fails", func(t *testing.T) {
-		_, _, err := tree.HonestPathTimer(ctx, id("blk-0.a-16.b"), blockNum)
+		_, err := tree.ComputeAncestorsWithTimers(ctx, id("blk-0.a-16.b"), blockNum)
 		require.ErrorContains(t, err, "not found in honest challenge tree")
 	})
 	t.Run("block challenge: level zero edge has no ancestors", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("blk-0.a-16.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("blk-0.a-16.a"), blockNum)
 		require.NoError(t, err)
-		require.Equal(t, 0, len(ancestors))
+		require.Equal(t, 0, len(resp.AncestorEdgeIds))
 	})
 	t.Run("block challenge: single ancestor", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("blk-0.a-8.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("blk-0.a-8.a"), blockNum)
 		require.NoError(t, err)
-		require.Equal(t, HonestAncestors{id("blk-0.a-16.a")}, ancestors)
-		_, ancestors, err = tree.HonestPathTimer(ctx, id("blk-8.a-16.a"), blockNum)
+		require.Equal(t, HonestAncestors{id("blk-0.a-16.a")}, resp.AncestorEdgeIds)
+		resp, err = tree.ComputeAncestorsWithTimers(ctx, id("blk-8.a-16.a"), blockNum)
 		require.NoError(t, err)
-		require.Equal(t, HonestAncestors{id("blk-0.a-16.a")}, ancestors)
+		require.Equal(t, HonestAncestors{id("blk-0.a-16.a")}, resp.AncestorEdgeIds)
 	})
 	t.Run("block challenge: many ancestors", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("blk-4.a-5.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("blk-4.a-5.a"), blockNum)
 		require.NoError(t, err)
 		wanted := HonestAncestors{
 			id("blk-4.a-6.a"),
@@ -89,10 +89,10 @@ func TestAncestors_AllChallengeLevels(t *testing.T) {
 			id("blk-0.a-8.a"),
 			id("blk-0.a-16.a"),
 		}
-		require.Equal(t, wanted, ancestors)
+		require.Equal(t, wanted, resp.AncestorEdgeIds)
 	})
 	t.Run("big step challenge: level zero edge has ancestors from block challenge", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("big-0.a-16.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("big-0.a-16.a"), blockNum)
 		require.NoError(t, err)
 		wanted := HonestAncestors{
 			id("blk-4.a-5.a"),
@@ -101,10 +101,10 @@ func TestAncestors_AllChallengeLevels(t *testing.T) {
 			id("blk-0.a-8.a"),
 			id("blk-0.a-16.a"),
 		}
-		require.Equal(t, wanted, ancestors)
+		require.Equal(t, wanted, resp.AncestorEdgeIds)
 	})
 	t.Run("big step challenge: many ancestors plus block challenge ancestors", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("big-5.a-6.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("big-5.a-6.a"), blockNum)
 		require.NoError(t, err)
 		wanted := HonestAncestors{
 			// Big step chal.
@@ -119,10 +119,10 @@ func TestAncestors_AllChallengeLevels(t *testing.T) {
 			id("blk-0.a-8.a"),
 			id("blk-0.a-16.a"),
 		}
-		require.Equal(t, wanted, ancestors)
+		require.Equal(t, wanted, resp.AncestorEdgeIds)
 	})
 	t.Run("small step challenge: level zero edge has ancestors from big and block challenge", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("smol-0.a-16.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("smol-0.a-16.a"), blockNum)
 		require.NoError(t, err)
 		wanted := HonestAncestors{
 			// Big step chal.
@@ -138,10 +138,10 @@ func TestAncestors_AllChallengeLevels(t *testing.T) {
 			id("blk-0.a-8.a"),
 			id("blk-0.a-16.a"),
 		}
-		require.Equal(t, wanted, ancestors)
+		require.Equal(t, wanted, resp.AncestorEdgeIds)
 	})
 	t.Run("small step challenge: lowest level edge has full ancestry", func(t *testing.T) {
-		_, ancestors, err := tree.HonestPathTimer(ctx, id("smol-5.a-6.a"), blockNum)
+		resp, err := tree.ComputeAncestorsWithTimers(ctx, id("smol-5.a-6.a"), blockNum)
 		require.NoError(t, err)
 		wanted := HonestAncestors{
 			// Small step chal.
@@ -162,7 +162,7 @@ func TestAncestors_AllChallengeLevels(t *testing.T) {
 			id("blk-0.a-8.a"),
 			id("blk-0.a-16.a"),
 		}
-		require.Equal(t, wanted, ancestors)
+		require.Equal(t, wanted, resp.AncestorEdgeIds)
 	})
 }
